Remove src/duckdb and merge into index
