+++ src/test_utils/mod.rs
@@ -3,21 +3,15 @@
 use std::sync::Arc;
 
 use anyhow::Result;
-use ratatui::{backend::TestBackend, Terminal};
-use swiftide::agents::tools::local_executor::LocalExecutor;
-use swiftide::agents::{Agent, DefaultContext};
-use swiftide::chat_completion::errors::LanguageModelError;
-use swiftide::chat_completion::{ChatCompletion, ChatCompletionResponse};
-use swiftide::traits::{EmbeddingModel, Persist as _, SimplePrompt, ToolExecutor};
-use tokio_util::task::AbortOnDropHandle;
-use uuid::Uuid;
+use swiftide::{
+    agents::{tools::local_executor::LocalExecutor, Agent, DefaultContext},
+    chat_completion::{errors::LanguageModelError, ChatCompletion, ChatCompletionResponse},
+    traits::{EmbeddingModel, SimplePrompt, ToolExecutor},
+};
 
-#[cfg(feature = "duckdb")]
-use crate::duckdb;
-use crate::frontend;
-#[cfg(feature = "duckdb")]
-use crate::indexing::duckdb_index::DuckdbIndex;
-use crate::{commands::CommandHandler, config::Config, frontend::App, git, repository::Repository};
+use crate::{config::Config, git, repository::Repository};
+
+pub mod integration;
 
 pub struct TestGuard {
     pub tempdir: tempfile::TempDir,
@@ -251,64 +245,6 @@ macro_rules! assert_agent_responded {
             .await
             .unwrap();
     };
-}
-
-pub struct IntegrationContext {
-    pub app: App<'static>,
-    pub uuid: Uuid,
-    pub repository: Arc<Repository>,
-    pub terminal: Terminal<TestBackend>,
-    pub workdir: std::path::PathBuf,
-
-    // Guards the command handler
-    pub handler_guard: AbortOnDropHandle<()>,
-    // Guards the repository
-    pub repository_guard: TestGuard,
-}
-
-impl IntegrationContext {
-    pub fn render_ui(&mut self) -> &TestBackend {
-        self.terminal
-            .draw(|f| frontend::ui(f, f.area(), &mut self.app))
-            .unwrap();
-
-        self.terminal.backend()
-    }
-}
-
-/// Sets up an app
-#[cfg(feature = "duckdb")]
-pub async fn setup_integration() -> Result<IntegrationContext> {
-    let (repository, repository_guard) = test_repository();
-    let workdir = repository.path().clone();
-    let repository = Arc::new(repository);
-    let mut app = App::default_from_repository(repository.clone()).with_workdir(repository.path());
-    let duckdb = duckdb::get_duckdb(&repository);
-    duckdb.setup().await.unwrap();
-    let terminal = Terminal::new(TestBackend::new(160, 40)).unwrap();
-
-    let index = DuckdbIndex::default();
-    let mut handler = CommandHandler::from_index(index);
-    handler.register_ui(&mut app);
-    let handler_guard = handler.start();
-
-    let uuid = Uuid::parse_str("a1a2a3a4b1b2c1c2d1d2d3d4d5d6d7d8").unwrap();
-    let current_chat = app.current_chat_mut();
-
-    // Force to fixed uuid so that snapshots are stable
-    current_chat.uuid = uuid;
-    app.current_chat_uuid = uuid;
-
-    Ok(IntegrationContext {
-        app,
-        uuid,
-        repository,
-        terminal,
-        workdir,
-
-        handler_guard,
-        repository_guard,
-    })
 }
 
 pub struct TempEnv<'a> {
